[VirtualKeyBoard.py] Disconnect function name from button
Naming a function based on one of the remote control buttons used to commonly activate the function is inappropriate.  This function can also be triggered by the ENTER key.

The other aspect of this correction is to reinforce that it is the keymap and not the code that creates associations between actual buttons and functions.