resolveFilename: do not change base path
Do not use a skin directory if it already exists in the base path.
This allows to keep the base path unchanged, as before, if nothing not found.

Also, add an additional path in the alternate directories search before the skin_default.
This allow check the path to a files in different skins, where they are specified.
For example if in skin PLi-FullNightHD is link to file in PLi-FullHD.
Before this path was not checked and returned default link if such image is not found in skin_default.