accel: fix subtitle rendering problem
The problem is caused by wrong data type that invalidates color entries in the palette in 64bit.

The size of int in C/C++ language is 32 bit on both 32 and 64 bit machine.
But, the size of long int is different: 32 bit on 32 bit machine while 64 bit on 64 bit machine.
So, int type must be used for 32-bit ARGB color value on both systems.

Signed-off-by: Athanasios Oikonomou <athoik@gmail.com>