TimerEdit: Solve GSOD for fallback timers when info button is pressed
The info buttons shows the timer log. The tomer log from a fallbacktuner
is currently not available so when we try to show the log for a
fallbacktuner timer we ignore it. With a change of the skin the info
button will also be automatically shown/hidden when the function is
available or not