Autoinstall: Solve the issue introduced during this day
It seems that the add opkg upgrade did lead to the fact nothing was
installed and the skipped already installed packages introduced wierd
races. Now when a package is already installed I just give an echo
command that the install is skipped which is also logged and shown on
the screen. The opkg upgrade is the first command given. Also add a 5
second timeout before the screen closes so you better can see the status
the autoinstall was completed.