FlashImage: Allow online flashing on network drives
Allow online flash on network drivers when no local drive such as an USB
or HDD is available. When a network drive is used auto restore cannot
work and that was why initially network drives were not available. When
a network drive is choosen the user will get a notification that the
backup itself can still be created on the network drive but to use it he
need first to (re)mount the network drive and then can run a
manufacturer reset to execute a restore when the user wants