When shutting down box also warn when client is streaming
When shutting down (deep standby) you get warned when a recording was
running or will soon be started and jobs were running. But not when a
client (fallbacktuner) is streaming from your box or watching recordings
from the harddisk drive. Now when shutting down the box you get also a
warning on these events. Also add the same warning to FlashImage as here
you also reboot the box.