Improve multituner support
Now when we have a multituner you can select between the tuners with
enable/disable and you do not have a seperate enable/disable anymore.

In addition, but not tested yet, when the tuner also has DVB-S you
should also be capable of selecting DVB-S + DVB-C (or DVB-T) at once. Then you
also can enter both parameters in that config screen. But as I do not
have such tuner I could not fully test this.

At least for a combi DVB-C/T it was tested

As I might need help here and it is limited to just a few boxes I did
commit this already in Develop fpr testing