Pixmap caching
C++ layer pixmap caching:
loadPNG and loadJPG have been updated to allow caching using an additional defaulted parameter.
loadPNG defaults to cached
loadJPG defaults to not cached
Caching is implemented in a helper class in pixmapcache.cpp. The cache holds a reference to the pixmap in the same manner as the previous caching. Images are reloaded from disk when the file modified time or the file size changes.
If the cache reaches a size of 256 images, the least recently referenced pixmap is removed. This is, I think, the same behaviour.
LoadPixmap has been updated to pass its cached parameter through or to use the default rules if it is not supplied by the caller.

I found that the cache caused protection fault crashes on exiting the enigma2 process when the cache held ePtr<gPixmap>, so I switched to using a gPixmap* and manually managing the refcounts. Since then, no crashes.