[Directories.py} Provide "#" separator workaround
This change is a temporary workaround to address an Enigma2 crash because the "#" character has been selected as a suffix separator for image paths.  All other images use ":" as this is often viewed as illegal because it *is* illegal in Windows and Apple environments.

When the skins and code that use "#" are corrected then this change *should* be reverted!