[ActionMap.py] Improve initialisation and logging
- Use a slightly longer form of initialisation for the ActionMap __init__. This format avoids an issue where the initialisation parameters become persistent across all invocations of this code. This is inappropriate and undesirable.

In essence the problem with the original code is that it appears to be an efficient way to create a default empty list and dictionary. What actually happens is that these items get created as a persistent objects. Every invocation of the __init__ method that doesn't specify the parameters will use that SAME (persistent) objects. Any changes to either object will persist (be remembered) and be carried forward into every other invocation of this method! THIS IS NOT WHAT SHOULD HAPPEN!

See  https://docs.python-guide.org/writing/gotchas/

- Add new logging code to report actions that are not accessible due to no matching keymap entries.  This may not be an error condition!

- Improve variable name clarity.

- Code is now common between main images.