Multiboot: Improve retreiving last upgrade date from image
Drive the last upgrade date from the image by checking the status file
from ipkg. This also should work for each image.

However when doing autoinstall/autorestore after a (re)flash the box
does not have a transponder time yet. Then it could happen that the
status file get 1970-01-01 as date. When this happens the date of the
image is taken from the bootlogo.mvi file and the enigma2 binary file
whichever is later. In addition when the bootlogo.mvi file does not
exists or a date cannot be retreived from it we do not get anymore a
GSOD.

Possibly for this work-a-round (status get date 1970-01-01 due to
autorestore/install) we need to find a different solution later.

# Conflicts:
#	lib/python/Tools/Multiboot.py